consolidate rabbitmq renderer
