+++ pkg/renderers/rabbitmqv1alpha3/renderer.go
@@ -14,16 +14,16 @@ import (
 	"github.com/project-radius/radius/pkg/renderers"
 )
 
-var _ renderers.Renderer = (*AzureRenderer)(nil)
+var _ renderers.Renderer = (*Renderer)(nil)
 
-type AzureRenderer struct {
+type Renderer struct {
 }
 
-func (r *AzureRenderer) GetDependencyIDs(ctx context.Context, workload renderers.RendererResource) ([]azresources.ResourceID, []azresources.ResourceID, error) {
+func (r *Renderer) GetDependencyIDs(ctx context.Context, workload renderers.RendererResource) ([]azresources.ResourceID, []azresources.ResourceID, error) {
 	return nil, nil, nil
 }
 
-func (r *AzureRenderer) Render(ctx context.Context, options renderers.RenderOptions) (renderers.RendererOutput, error) {
+func (r *Renderer) Render(ctx context.Context, options renderers.RenderOptions) (renderers.RendererOutput, error) {
 	properties := &radclient.RabbitMQMessageQueueResourceProperties{}
 	resource := options.Resource
 	err := resource.ConvertDefinition(&properties)
